added the profile interface. moved the storages to their own dir in preparation for adding more flexibility to it.
