+++ src/Symfony/Component/Profiler/DataCollector/EventDataCollector.php
@@ -15,6 +15,7 @@ use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 use Symfony\Component\EventDispatcher\Debug\TraceableEventDispatcherInterface;
 use Symfony\Component\Profiler\Context\ContextInterface;
 use Symfony\Component\Profiler\Profile;
+use Symfony\Component\Profiler\ProfileInterface;
 
 /**
  * EventDataCollector.
@@ -42,7 +43,7 @@ class EventDataCollector extends DataCollector implements LateDataCollectorInter
         return true;
     }
 
-    public function lateCollect()
+    public function lateCollect(ProfileInterface $profile)
     {
         if ($this->dispatcher instanceof TraceableEventDispatcherInterface) {
             $this->setCalledListeners($this->dispatcher->getCalledListeners());
