+++ src/Symfony/Component/Profiler/DataCollector/MemoryDataCollector.php
@@ -13,6 +13,7 @@ namespace Symfony\Component\Profiler\DataCollector;
 
 use Symfony\Component\Profiler\Context\ContextInterface;
 use Symfony\Component\Profiler\Profile;
+use Symfony\Component\Profiler\ProfileInterface;
 
 /**
  * MemoryDataCollector.
@@ -41,7 +42,7 @@ class MemoryDataCollector extends DataCollector implements LateDataCollectorInte
     /**
      * {@inheritdoc}
      */
-    public function lateCollect()
+    public function lateCollect(ProfileInterface $profile)
     {
         $this->updateMemoryUsage();
     }
