+++ src/Symfony/Component/Profiler/Storage/FileProfilerStorage.php
@@ -9,7 +9,10 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Component\Profiler;
+namespace Symfony\Component\Profiler\Storage;
+
+use Symfony\Component\Profiler\Profile;
+use Symfony\Component\Profiler\Storage\ProfilerStorageInterface;
 
 /**
  * Storage for profiler using files.
